backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.0

# AI/LLM integration
openai==1.10.0
anthropic==0.8.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.26.0

# CORS
python-jose[cryptography]==3.3.0
